NFC: Encapsulate `Budget` internals.
Make it clear that nothing is reaching into the internals of `Budget`.
In particular, clarify that the tests are not messing around with the
defaults.
